Feat Solves #9 Create branch `assessment` with Gradle
Created a project with Gradle with :
* one Main that is called `HelloWorld` that just say `HelloWorld` and a method `returnOne` that returns 1 so that we have a method we can test.
* one test class `HelloWorldTest` that has a test that check if
`returnOne` returns 1.

The project has a gradle so that we can use `gradle build` and `gradle
clean test` to build and run the tests.
Then the result of the tests will be in the
`build/test-results/test/TEST-HelloWorldTest.xml` that we will be able
to use to give feedback to the user.